Moved storage helper to storage folder
