+++ src/options/Options.tsx
@@ -1,7 +1,7 @@
 import { acquireVideoStream } from '@src/helpers/optionsActions';
 import React, { useEffect, useRef, useState } from 'react';
 import { useChromeStorageLocal } from '@eamonwoortman/use-chrome-storage';
-import { removeFromStorage } from '@src/helpers/storage';
+import { removeFromStorage } from '@src/storage/storage';
 import '../css/app.css';
 import "./styles.module.css";
